refactor(archtest): TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01 迁 module-path-agnostic + 缩 baseline [#1639]
移 Check* + scanner helper 到非 test eval_predicate_centralization.go（2 个
typeseval/archtest pkg 路径派生自 PlatformModulePath），_test.go 瘦身为 dogfood
+ RED fixture 单源。register=none（framework-self-referential meta-rule，扫
archtest 自身 _test.go，外部 vacuous）。baseline 删该行。
