+++ tools/archtest/eval_predicate_centralization_test.go
@@ -1,334 +1,27 @@
-package archtest
-
 // INVARIANT: TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01
 //
-// eval_predicate_centralization_test.go — every go/build/constraint.Expr.Eval
-// callsite in tools/archtest/*_test.go (top-level package, excluding
-// internal/ subpackages) must pass either typeseval.BuildContextPredicate(...)
-// (or equivalently the archtest facade BuildContextPredicate(...) which is a
-// thin delegation to the same function) or an inline
-// `func(_ string) bool { return false }` sentinel. Hand-written tag predicates
-// drift past go-toolchain default tag additions; this funnel makes the wrong
-// shape archtest-detectable.
-//
-// Single-rule file per ai-robust.md "archtest 文件命名" branch (single rule →
-// {rule}_test.go). Promote to {theme}_invariants_test.go if related TYPESEVAL-*
-// invariants accumulate to ≥ 3.
+// This _test.go only dogfoods + precision-gates the rule. The importable rule
+// body (CheckEvalPredicateCentralization01 + scanFileForEvalPredicateViolations
+// + isConstraintExprEvalCall + isBuildContextPredicateCallee +
+// isAllFalseSentinelFuncLit + formatEvalCallee + evalPredicateViolation + the
+// full package godoc, all module-path-agnostic via PlatformModulePath) lives in
+// the non-test companion eval_predicate_centralization.go (M3 #1639) so it is
+// fork-safe — single source, no parallel rule body.
+package archtest
 
 import (
-	"go/ast"
-	"go/token"
-	"go/types"
 	"path/filepath"
-	"sort"
-	"strings"
 	"testing"
 )
 
-const (
-	// buildContextPredicatePkg is the canonical typeseval package; imports from
-	// external packages (outside tools/archtest) must use this qualified form.
-	buildContextPredicatePkg = "github.com/ghbvf/gocell/tools/archtest/internal/typeseval"
-	// buildContextPredicateFacadePkg is the archtest package itself — same-package
-	// test files (package archtest) call the unqualified facade wrapper
-	// BuildContextPredicate which resolves here, not to typeseval. Both are
-	// accepted canonical forms because the facade is a thin delegation.
-	buildContextPredicateFacadePkg = "github.com/ghbvf/gocell/tools/archtest"
-	buildContextPredicateFunc      = "BuildContextPredicate"
-	constraintExprPkgPath          = "go/build/constraint"
-	constraintExprEvalMethod       = "Eval"
-)
-
-// INVARIANT: TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01
-//
-// Every constraint.Expr.Eval(arg) callsite in tools/archtest/*_test.go (top-
-// level package, excluding internal/ subpackages) MUST pass either:
-//
-//	Form A — a *ast.CallExpr whose callee resolves to either
-//	         tools/archtest/internal/typeseval.BuildContextPredicate
-//	         (qualified cross-package form) OR
-//	         tools/archtest.BuildContextPredicate (same-package bare form,
-//	         which is a thin delegation to typeseval.BuildContextPredicate).
-//	         Any extraTags arguments are fine; the funnel is the call form.
-//	Form B — an inline *ast.FuncLit of shape  func(_ string) bool { return false }
-//	         — the all-false sentinel — body must be exactly one ReturnStmt
-//	         whose single result is the identifier `false`. Indirection via a
-//	         var binding (`var fn = func(...) { return false }; expr.Eval(fn)`)
-//	         is NOT accepted: the inline FuncLit shape is the Hard surface.
-//
-// Any other predicate shape — `func(tag string) bool { return tag == "X" }`,
-// a named helper, a variable identifier, a method reference, a composite
-// return expression — fails this rule.
-//
-// Why a static funnel (Hard) and not a doc convention (Soft):
-//
-//   - The toolchain-default tag set (GOOS/GOARCH/cgo/unix/gc/go1.X) drifts
-//     every release. A hand-written predicate that hard-codes "linux ||
-//     darwin || amd64 || ..." silently misses additions. BuildContextPredicate
-//     sources implicitDefaults from build.Default.ReleaseTags + a single
-//     mirror of internal/syslist, so go.mod floor bumps propagate
-//     automatically. Forcing every consumer through the call form makes the
-//     drift unreachable.
-//
-//   - The all-false sentinel is the canonical "evaluate under empty tag
-//     set" form; allowing it explicitly avoids forcing a contrived
-//     BuildContextPredicate() call where the intent is to deny every tag.
-//
-// Implementation parallels PRODUCTION-LOADER-FUNNEL-01 (ban one shape,
-// allowlist by structural exclusion) and PANIC-REGISTERED-01 (form
-// uniqueness at panic() callsite).
-//
-// AI-robust rating: Hard. Per .claude/rules/gocell/ai-robust.md "Hard 范本"
-// — "typed function call as Hard funnel for unbounded operations": form
-// uniqueness + archtest fail-on-deviation is the highest grade reachable in
-// Go for this rule shape. The Go type system does not prevent passing
-// arbitrary func(string)bool to constraint.Expr.Eval at compile time;
-// static archtest at CI is the canonical Hard enforcement.
-//
-// Internal scope note: tools/archtest/internal/typeseval/buildtags_test.go
-// is the implementor's coverage self-test for BuildContextPredicate's flat
-// load semantics; it uses derived per-slice predicates by design and is
-// outside this rule's scope (mirroring panic_invariants excluding
-// pkg/panicregister/*_test.go from PANIC-REGISTERED-01).
+// TestEvalPredicateCentralization01 dogfoods
+// TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01 against GoCell itself by calling the
+// same CheckEvalPredicateCentralization01 that holds the importable rule body —
+// single source. Every constraint.Expr.Eval callsite in
+// tools/archtest/*_test.go must pass either typeseval.BuildContextPredicate(...)
+// (Form A) or an inline `func(_ string) bool { return false }` sentinel (Form B).
 func TestEvalPredicateCentralization01(t *testing.T) {
-	t.Parallel()
-	root := findModuleRoot(t)
-
-	// Run(t, Typed(TypedOpts{Tests: true}, ...)) loads *_test.go files so the rule can
-	// inspect archtest tests themselves. Pattern is scoped to ./tools/archtest/...
-	// — not the real-repo "./..." that PRODUCTION-LOADER-FUNNEL-01 polices —
-	// so this is the canonical narrow-scope archtest-self load shape, same
-	// as scanner_framework_usage_test.go.
-	var violations []evalPredicateViolation
-	Run(t, Typed(TypedOpts{Tests: true}, []string{"./tools/archtest/..."}),
-		func(p *Pass) []Diagnostic {
-			for _, file := range p.Files {
-				rel := p.Rel(file)
-
-				if filepath.ToSlash(filepath.Dir(rel)) != "tools/archtest" {
-					continue
-				}
-				if !strings.HasSuffix(rel, "_test.go") {
-					continue
-				}
-				violations = append(violations,
-					scanFileForEvalPredicateViolations(p.Fset, file, p.TypesInfo, rel)...)
-			}
-			return nil
-		})
-
-	_ = root
-
-	sort.Slice(violations, func(i, j int) bool {
-		if violations[i].Rel != violations[j].Rel {
-			return violations[i].Rel < violations[j].Rel
-		}
-		return violations[i].Line < violations[j].Line
-	})
-	for _, v := range violations {
-		t.Errorf("TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01: %s:%d: "+
-			"constraint.Expr.Eval predicate must be typeseval.BuildContextPredicate(...) "+
-			"or inline func(_ string) bool { return false }; got %s. "+
-			"Reason: the toolchain-default tag set (GOOS/GOARCH/cgo/unix/gc/go1.X) "+
-			"drifts every release; hand-written predicates miss additions. Use "+
-			"typeseval.BuildContextPredicate(extraTags...) to inherit defaults.",
-			v.Rel, v.Line, v.Form)
-	}
-}
-
-// evalPredicateViolation records one TYPESEVAL-EVAL-PREDICATE-CENTRALIZED-01
-// hit for batched reporting.
-type evalPredicateViolation struct {
-	Rel  string
-	Line int
-	Form string // human-readable description of the offending shape
-}
-
-// scanFileForEvalPredicateViolations walks file's AST for calls shaped
-// `<expr>.Eval(arg)` whose receiver static type is go/build/constraint.Expr,
-// and records a violation for each call whose first argument matches neither
-// Form A (typeseval.BuildContextPredicate(...) CallExpr) nor Form B (inline
-// all-false sentinel FuncLit).
-func scanFileForEvalPredicateViolations(
-	fset *token.FileSet,
-	file *ast.File,
-	info *types.Info,
-	rel string,
-) []evalPredicateViolation {
-	var violations []evalPredicateViolation
-
-	EachInSubtree[ast.CallExpr](file, func(call *ast.CallExpr) {
-		if !isConstraintExprEvalCall(call, info) {
-			return
-		}
-		if len(call.Args) == 0 {
-			violations = append(violations, evalPredicateViolation{
-				Rel:  rel,
-				Line: fset.Position(call.Pos()).Line,
-				Form: "<no arguments>",
-			})
-			return
-		}
-		arg := call.Args[0]
-
-		// Form A: arg is a CallExpr whose callee resolves to
-		// typeseval.BuildContextPredicate.
-		if argCall, ok := arg.(*ast.CallExpr); ok {
-			if isBuildContextPredicateCallee(argCall.Fun, info) {
-				return
-			}
-			violations = append(violations, evalPredicateViolation{
-				Rel:  rel,
-				Line: fset.Position(call.Pos()).Line,
-				Form: "CallExpr with non-canonical callee " + formatEvalCallee(argCall.Fun),
-			})
-			return
-		}
-
-		// Form B: arg is an inline FuncLit shaped func(_ string) bool { return false }.
-		if fl, ok := arg.(*ast.FuncLit); ok {
-			if isAllFalseSentinelFuncLit(fl) {
-				return
-			}
-			violations = append(violations, evalPredicateViolation{
-				Rel:  rel,
-				Line: fset.Position(call.Pos()).Line,
-				Form: "FuncLit with non-canonical body (not `return false` single-stmt)",
-			})
-			return
-		}
-
-		// Any other shape (Ident, SelectorExpr, etc.) — including indirection
-		// through a var-bound predicate — is a violation. The inline FuncLit
-		// requirement is what makes Form B Hard.
-		violations = append(violations, evalPredicateViolation{
-			Rel:  rel,
-			Line: fset.Position(call.Pos()).Line,
-			Form: "non-CallExpr/non-FuncLit predicate argument (var binding / Ident / SelectorExpr not allowed)",
-		})
-	})
-
-	return violations
-}
-
-// isConstraintExprEvalCall reports whether call is `<expr>.Eval(...)` where
-// the resolved method is go/build/constraint.Expr.Eval. Uses types.Info
-// Selections (via typeseval.ResolveMethodCall) to recover the method's owning
-// package and name. When info is nil (fixture / partial type info) the
-// detection falls through to false — this is fail-closed for the
-// caller-visible direction (we never over-flag a non-constraint Eval; the
-// archtest may under-detect on missing type info, which manifests as the
-// load failing earlier via require.NoError).
-func isConstraintExprEvalCall(call *ast.CallExpr, info *types.Info) bool {
-	if info == nil {
-		return false
-	}
-	sel, ok := call.Fun.(*ast.SelectorExpr)
-	if !ok || sel.Sel == nil {
-		return false
-	}
-	if sel.Sel.Name != constraintExprEvalMethod {
-		return false
-	}
-	fn, ok := ResolveMethodCall(info, sel)
-	if !ok {
-		return false
-	}
-	if fn.Pkg() == nil {
-		return false
-	}
-	return fn.Pkg().Path() == constraintExprPkgPath && fn.Name() == constraintExprEvalMethod
-}
-
-// isBuildContextPredicateCallee reports whether funExpr refers to either
-// typeseval.BuildContextPredicate (qualified, from external packages) or the
-// archtest facade BuildContextPredicate (bare Ident, from same-package test
-// files in package archtest). Uses types.Info via typeseval.ResolvePackageRef
-// (handles both qualified `typeseval.X` and dot-imported bare `X` / same-pkg
-// bare `X` forms); falls back to AST-only `typeseval.X` selector matching when
-// info is nil. Mirrors panic_invariants_test.go::isApprovedCallee structure.
-//
-// Both accepted packages are canonical: buildContextPredicateFacadePkg wraps
-// buildContextPredicatePkg as a thin delegation (archtest.BuildContextPredicate
-// → typeseval.BuildContextPredicate). Same-package test files call the facade
-// form and both resolve to the same semantic guarantee.
-func isBuildContextPredicateCallee(funExpr ast.Expr, info *types.Info) bool {
-	isCanonicalPkg := func(pkgPath string) bool {
-		return pkgPath == buildContextPredicatePkg || pkgPath == buildContextPredicateFacadePkg
-	}
-
-	sel, ok := funExpr.(*ast.SelectorExpr)
-	if !ok || sel.Sel == nil {
-		// Bare Ident is also possible under dot-import or same-package call;
-		// ResolvePackageRef handles both *ast.SelectorExpr and *ast.Ident.
-		if info != nil {
-			if pkgPath, name, ok := ResolvePackageRef(info, funExpr); ok {
-				return isCanonicalPkg(pkgPath) && name == buildContextPredicateFunc
-			}
-		}
-		return false
-	}
-	if sel.Sel.Name != buildContextPredicateFunc {
-		return false
-	}
-	if info != nil {
-		if pkgPath, name, ok := ResolvePackageRef(info, sel); ok {
-			return isCanonicalPkg(pkgPath) && name == buildContextPredicateFunc
-		}
-		// info is non-nil but ResolvePackageRef declined: sel.X is not a
-		// package qualifier (e.g. method-position selector on a value with
-		// a `BuildContextPredicate` method). Decline rather than falling
-		// through to the AST-only path — that path matches any Ident named
-		// "typeseval" and would over-accept under method-position shadowing.
-		return false
-	}
-	// AST-only fallback (no info): match `typeseval.BuildContextPredicate`.
-	xIdent, ok := sel.X.(*ast.Ident)
-	if !ok {
-		return false
-	}
-	return xIdent.Name == "typeseval"
-}
-
-// isAllFalseSentinelFuncLit reports whether fl is exactly:
-//
-//	func(<any-ident>) bool { return false }
-//
-// — body has exactly 1 statement, that statement is a ReturnStmt with
-// exactly 1 result, and the result is the *ast.Ident `false`. Parameter
-// and result types are not re-checked: the constraint.Expr.Eval signature
-// already constrains them to func(string)bool at the type system.
-//
-// Equivalent forms like `return !true`, `return 0 == 1`, or a var binding
-// (`var f = func(...) bool { return false }; expr.Eval(f)`) are NOT
-// accepted. The inline-literal-with-`false`-ident shape is the Hard
-// surface — anything else fails archtest.
-//
-// Note on `false` shadowing: Go permits redefining the predeclared
-// `false` identifier via assignment (e.g. `false := true`), but only
-// across multiple statements. The single-statement body check
-// (len(fl.Body.List) == 1) structurally rules out a shadowing pattern,
-// so `id.Name == "false"` is a complete check — there is no
-// `{ false := true; return false }` form that fits in 1 statement.
-func isAllFalseSentinelFuncLit(fl *ast.FuncLit) bool {
-	if fl == nil || fl.Body == nil {
-		return false
-	}
-	if len(fl.Body.List) != 1 {
-		return false
-	}
-	ret, ok := fl.Body.List[0].(*ast.ReturnStmt)
-	if !ok {
-		return false
-	}
-	if len(ret.Results) != 1 {
-		return false
-	}
-	id, ok := ret.Results[0].(*ast.Ident)
-	if !ok {
-		return false
-	}
-	return id.Name == "false"
+	Report(t, evalPredicateRuleID, CheckEvalPredicateCentralization01(t, ConfigForExternalCell{}))
 }
 
 // TestEvalPredicateCentralizationFixtures is the "test the test" meta-test
@@ -391,22 +84,4 @@ func TestEvalPredicateCentralizationFixtures(t *testing.T) {
 			AssertGolden(t, filepath.Join(fixtureDir, "diag.golden"), diags)
 		})
 	}
-}
-
-// formatEvalCallee returns a short human-readable callee for violation
-// messages. Mirrors panic_invariants_test.go::formatCallee.
-func formatEvalCallee(funExpr ast.Expr) string {
-	switch fun := funExpr.(type) {
-	case *ast.Ident:
-		return fun.Name
-	case *ast.SelectorExpr:
-		if fun.Sel == nil {
-			return "<unknown>"
-		}
-		if xIdent, ok := fun.X.(*ast.Ident); ok {
-			return xIdent.Name + "." + fun.Sel.Name
-		}
-		return "?." + fun.Sel.Name
-	}
-	return "<unknown>"
 }
